LoggerFactory: inject plugin managers. PsrLoggerAbstractAdapterFactory now composes LoggerAbstractServiceFactory
